+++ packages/react-query/src/__tests__/utils.tsx
@@ -1,7 +1,7 @@
 import { vi } from 'vitest'
 import * as React from 'react'
 import { act, render } from '@testing-library/react'
-import { environmentManager, isServer } from '@tanstack/query-core'
+import { environmentManager } from '@tanstack/query-core'
 import { QueryClientProvider, onlineManager } from '..'
 import type { QueryClient } from '..'
 import type { MockInstance } from 'vitest'
@@ -57,8 +57,9 @@ export function setActTimeout(fn: () => void, ms?: number) {
 }
 
 export function setIsServer(value: boolean) {
+  const originalIsServer = environmentManager.isServer()
   environmentManager.setIsServer(() => value)
   return () => {
-    environmentManager.setIsServer(() => isServer)
+    environmentManager.setIsServer(() => originalIsServer)
   }
 }